Reajuste_final
 - LoginUiActivity otimizado.
 - LoginEmailActivity otimizado.
 - ViewCadastroActivity otimizado.
 - CadastroUserActivity otimizado.
 - CadastroEmailTermosActivity         otimizado.
 - ProblemasLogin otimizado.
 - PermissaoSegundoPlanoActivity otimizado
 - FotoPerfilActivity otimizado.
 - AlterarSenhaActivity otimizado.

 -* Activitys
 - QRCodeActivity otimizado.
 - OfflineActivity otimizado.
 - FindPeopleActivity otimizado.
 - ConfigurarFotoActivity otimizado.

 -* Fragments
  - PhotoPostFragment otimizado.
  - GifPostFragment otimizado.
  - TextPostFragment otimizado.
  - VideoPostFragment otimizado.

  * Activitys e fragments anteriores otimizados e traduzidos.

 FollowersAndFollowingActivity  concluído, falta ajustes na dimensão do spinkit para cobrir toda o campo da view e em todos os searchview tem que organizar uma lógica para pesquisar usuários por nome sem custos desnecessários, FriendInteractionsActivity concluído, traduzido e com operações atomicas com uso de hashmap e updatechildren falta elaborar essa mesma abordagem em FollowersAndFollowingActivity.

   * Arrumado um pequeno erro de lógica em FindPeopleActivity, se o usuário pesquisado tivesse o id igual ao usuário atual ele iria ignorar os outros usuários depois do atual, removido o else dessa lógica o erro foi arrumado.

   e em fragments o aconselhado é usar onDestroyView para liberar o que quer que seja pois vem antes do onDestroy.

   * Ordenação de seguidores e seguindo agora são pelo timestampinteracao.

  * CommonPosting e PostUtils que são relacionadas aos fragments responsáveis pela criação e edição de postagens foram ajustados para utilizar hashmap com updatechildren, também foi feito uma pequena mudança no DetalhesPostagemActivity e AdapterFuncoesPostagem agora ele é capaz de notificar a mudança de interesses da postagem.

  * FindPeopleActivity e AdapterFindPeoples foram feitas pequenas alterações. Agora se o usuário atual está bloqueado pelo usuário listado, não será possível visitar o perfil do usuário e caso esse bloqueio foi antes da listagem a foto e o fundo do usuário listado será oculto.

  * AdapterFriends, AdapterRequests agora estão configurados para permitir somente funções específicas somente se o usuário atual não estiver sido bloqueado pelo usuário alvo.

  * AdapterFoll, PersonProfileActivity e SeguindoUtils foram modificadas para funcionar com hashmap e updatechildren, funcionando bem o seguir e o deixar de seguir, agora tem um boolean para controlar as operações no PersonProfileActivity também e também agora existe um listener para acompanhar o block do dono do perfil no usuário visitante onde no momento que o visitante for bloqueado ele será expulso do perfil do usuário.

  * Criação e edição de comunidade concluído contendo um listener para alterações nas comunidas onde o usuário atual é o dono, faltando apenas alguns ajustes na ListaComunidadesActivityNEW.

  * CommunityDetailsActivity em questão de funcionamento está quase concluído, detalhes estão sendo exibidos corretamente, é necessário criar uma lógica nas regras que impede de que mais de um usuário tente gerenciar a comunidade ao mesmo tempo, da para se basear na lógica do unlockUnfriend. O gerenciamento de usuários está funcionando bem, falta alguns testes e ajustes. Atualmente todas funções estão concluídas, falta ajustes também na parte de bloqueio e denúncia de comunidade e ainda tradução e ajustes nos layouts.

  * Lógica para exibição das comunidades de acordo com cada tipo de comunidade foi concluído, faltando apenas tradução, sistema de bloqueio onde o fundador ou adm pode bloquear um usuário / sistema de denúnciar comunidade, falta ajustes nos layouts e por fim pequenas otimizações.

  * Começando a otimizar a listagem de chats, contatos e chat em grupo.
  - Listagem de chats ainda em andamento, alteração e exclusão de dados tanto para filtragem como para sem filtragem está funcionando bem, a lógica de alterações remoções foi corrigido mais alguns erros. É necessário testar com mais dados pois o teste só foi testado com apenas 2 dados, listener duplicados arrumados e search funcionando pelo submit, falta ajustes mas todo bug até agora foi corrigido.

  * Ainda em teste mas já foi feito alterações para que sempre os chats mais recentes sejam adicionado no topo e se o usuário estiver observando os 3 primeiros itens, o usuário será automaticamente focado no primeiro elemento, falta verificar a troca de todas posições de cada maneira possível para ter certeza se está funcionando bem ou não.

  - Arrumado ou problema onde o viewpager apenas guardava na memória no máximo 2 fragments, isso foi mudado para 3 fragments agora o primeiro fragment não será destruído quando eu for para as outras abas do viewpager.

  * Parte da listagem de contatos já está pronta, falta alguns testes e implementações em torno do search e estará com o funcionamento concluído.

  obs: Falta fazer o teste em todas telas que envolve número.

- 16/03/2024 - 11:31

+++ app/src/main/java/com/example/ogima/helper/FriendsUtils.java
@@ -99,7 +99,7 @@ public class FriendsUtils {
     }
 
     public interface AjustarContatoCallback {
-        void onAjustar(String nivelAmizade, int totalMensagens);
+        void onAjustar(String nivelAmizade, long totalMensagens);
 
         void onError(String message);
     }
@@ -509,12 +509,12 @@ public class FriendsUtils {
 
         ajustarAmizadeEMsgContato(contadorMensagemRef, new AjustarContatoCallback() {
             @Override
-            public void onAjustar(String nivelAmizade, int totalMensagens) {
+            public void onAjustar(String nivelAmizade, long totalMensagens) {
                 operacoes.put(caminhoContatoAtual + "nivelAmizade", nivelAmizade);
                 operacoes.put(caminhoContatoAtual + "totalMensagens", totalMensagens);
                 ajustarAmizadeEMsgContato(contadorMensagemSelecionadoRef, new AjustarContatoCallback() {
                     @Override
-                    public void onAjustar(String nivelAmizade, int totalMensagens) {
+                    public void onAjustar(String nivelAmizade, long totalMensagens) {
                         operacoes.put(caminhoContatoAlvo + "nivelAmizade", nivelAmizade);
                         operacoes.put(caminhoContatoAlvo + "totalMensagens", totalMensagens);
                         firebaseRef.updateChildren(operacoes, new DatabaseReference.CompletionListener() {
